# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Central Limit Order Book (CLOB) with matching engine for trading crypto assets",
        "title": "Crypto Exchange API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/accounts/balance": {
            "get": {
                "description": "Get all balances for a user's account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Get account balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/credit": {
            "post": {
                "description": "Add balance to a user's account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Credit asset to account",
                "parameters": [
                    {
                        "description": "Credit details (includes user_id)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreditDebitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Credit successful",
                        "schema": {
                            "$ref": "#/definitions/v1.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/debit": {
            "post": {
                "description": "Remove balance from a user's account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Debit asset from account",
                "parameters": [
                    {
                        "description": "Debit details (includes user_id)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreditDebitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Debit successful",
                        "schema": {
                            "$ref": "#/definitions/v1.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orderbook": {
            "get": {
                "description": "Get the current orderbook for a trading pair",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orderbook"
                ],
                "summary": "Get orderbook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trading pair (e.g., BTC/BRL)",
                        "name": "pair",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orderbook retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.OrderbookResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Orderbook not found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orders": {
            "post": {
                "description": "Create a limit or market order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Place a new order",
                "parameters": [
                    {
                        "description": "Order details",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.PlaceOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order placed successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.PlaceOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/cancel": {
            "post": {
                "description": "Cancel an existing order by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "parameters": [
                    {
                        "description": "Cancel order details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CancelOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the health status of the API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "$ref": "#/definitions/v1.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.BalanceItem": {
            "type": "object",
            "properties": {
                "asset": {
                    "type": "string"
                },
                "available": {
                    "type": "number"
                },
                "locked": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "v1.BalanceResponse": {
            "type": "object",
            "properties": {
                "balances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.BalanceItem"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "v1.CancelOrderRequest": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer"
                },
                "pair": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "v1.CreditDebitRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "asset": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "v1.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "v1.LimitLevel": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "total_volume": {
                    "type": "number"
                }
            }
        },
        "v1.MatchResponse": {
            "type": "object",
            "properties": {
                "ask_order_id": {
                    "type": "integer"
                },
                "bid_order_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "size_filled": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "v1.OrderResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "filled_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "pair": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "side": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "v1.OrderbookResponse": {
            "type": "object",
            "properties": {
                "ask_total_volume": {
                    "type": "number"
                },
                "asks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.LimitLevel"
                    }
                },
                "bid_total_volume": {
                    "type": "number"
                },
                "bids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.LimitLevel"
                    }
                },
                "pair": {
                    "type": "string"
                },
                "spread": {
                    "type": "number"
                }
            }
        },
        "v1.PlaceOrderRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "pair": {
                    "description": "\"BTC/BRL\"",
                    "type": "string"
                },
                "price": {
                    "description": "0 para market orders",
                    "type": "number"
                },
                "side": {
                    "description": "\"bid\" ou \"ask\"",
                    "type": "string"
                },
                "type": {
                    "description": "\"limit\" ou \"market\"",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "v1.PlaceOrderResponse": {
            "type": "object",
            "properties": {
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.MatchResponse"
                    }
                },
                "order": {
                    "$ref": "#/definitions/v1.OrderResponse"
                }
            }
        }
    }
}